Finishes off new Sharing code
It's mostly the same code,
 but at least it's now easier to make changes to it
