+++ static/main.js
@@ -34,7 +34,6 @@ require('bootstrap-slider');
 
 var Sharing = require('./sharing').Sharing;
 var _ = require('underscore');
-var cloneDeep = require('lodash.clonedeep');
 var $ = require('jquery');
 var GoldenLayout = require('golden-layout');
 var Components = require('./components');
@@ -49,7 +48,6 @@ var themer = require('./themes');
 var motd = require('./motd');
 var jsCookie = require('js-cookie');
 var SimpleCook = require('./simplecook');
-var History = require('./history');
 var HistoryWidget = require('./history-widget').HistoryWidget;
 var presentation = require('./presentation');
 
@@ -225,6 +223,37 @@ function setupButtons(options) {
     }
 }
 
+function configFromEmbedded(embeddedUrl) {
+    // Old-style link?
+    var params;
+    try {
+        params = url.unrisonify(embeddedUrl);
+    } catch (e) {
+        // Ignore this, it's not a problem
+    }
+    if (params && params.source && params.compiler) {
+        var filters = _.chain((params.filters || '').split(','))
+            .map(function (o) {
+                return [o, true];
+            })
+            .object()
+            .value();
+        return {
+            content: [
+                {
+                    type: 'row',
+                    content: [
+                        Components.getEditorWith(1, params.source, filters),
+                        Components.getCompilerWith(1, filters, params.options, params.compiler),
+                    ],
+                },
+            ],
+        };
+    } else {
+        return url.deserialiseState(embeddedUrl);
+    }
+}
+
 function findConfig(defaultConfig, options) {
     var config;
     if (!options.embedded) {
@@ -260,7 +289,7 @@ function findConfig(defaultConfig, options) {
                 showCloseIcon: false,
                 hasHeaders: false,
             },
-        }, Sharing.configFromEmbedded(window.location.hash.substr(1)));
+        }, configFromEmbedded(window.location.hash.substr(1)));
     }
     return config;
 }
@@ -268,8 +297,9 @@ function findConfig(defaultConfig, options) {
 function initializeResetLayoutLink() {
     var currentUrl = document.URL;
     if (currentUrl.includes('/z/')) {
-        $('#ui-brokenlink').attr('href', currentUrl.replace('/z/', '/resetlayout/'));
-        $('#ui-brokenlink').show();
+        $('#ui-brokenlink')
+            .attr('href', currentUrl.replace('/z/', '/resetlayout/'))
+            .show();
     } else {
         $('#ui-brokenlink').hide();
     }
@@ -339,26 +369,6 @@ function initPolicies(options) {
             analytics.initialise();
         }
     }
-}
-
-function filterComponentState(config, keysToRemove) {
-    function filterComponentStateImpl(component) {
-        if (component.content) {
-            for (var i = 0; i < component.content.length; i++) {
-                filterComponentStateImpl(component.content[i], keysToRemove);
-            }
-        }
-
-        if (component.componentState) {
-            Object.keys(component.componentState)
-                .filter(function (key) { return keysToRemove.includes(key); })
-                .forEach(function (key) { delete component.componentState[key]; });
-        }
-    }
-
-    config = cloneDeep(config);
-    filterComponentStateImpl(config);
-    return config;
 }
 
 /*
@@ -476,10 +486,6 @@ function start() {
         layout = new GoldenLayout(defaultConfig, root);
         hub = new Hub(layout, subLangId, defaultLangId);
     }
-
-    var lastState = null;
-
-
 
     function sizeRoot() {
         var height = $(window).height() - (root.position().top || 0) - ($('#simplecook:visible').height() || 0);
